Move user classes to a new package
