+++ app/src/main/java/com/models/UserDatabase.java
@@ -6,7 +6,7 @@ import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.presenters.User;
+import com.presenters.users.User;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
